Finally got the system update disk file format fully understood
This commit contains code that successfully decodes and reencodes them
resulting in byte-identical files.

The code is a bit messy, so I added FIXME comments to make it clear that
the good decoding and encoding routines are currently on the wrong
script, called "kn5000_patch.py"

I will soon fix this, but I wanted to make the code public immediately.