work in progress
With this commit I am able to send some chars to an arduino connected to
the DB-15 parallel port of the HDAE5000 board

The code runs during boot, injected at the address of the "user did not request
firmware update" routine.

I have built a small preamble of assembly code to resume execution
cleanly after we return from our custom injected code. But something is
not right, so the keyboard freezes at that point. This needs further
debugging.